+++ 404.php
@@ -8,43 +8,57 @@
 get_header(); ?>
 
 	<div id="primary" class="content-area">
+
 		<main id="main" class="site-main" role="main">
 
 			<section class="error-404 not-found">
-				<header class="page-header">
-					<h1 class="page-title"><?php _e( 'Oops! That page can&rsquo;t be found.', 'storefront' ); ?></h1>
-				</header><!-- .page-header -->
 
 				<div class="page-content">
-					<p><?php _e( 'It looks like nothing was found at this location. Maybe try one of the links below or a search?', 'storefront' ); ?></p>
 
-					<?php get_search_form(); ?>
+					<header class="page-header">
+						<h1 class="page-title"><?php esc_html_e( 'Oops! That page can&rsquo;t be found.', 'storefront' ); ?></h1>
+					</header><!-- .page-header -->
 
-					<?php the_widget( 'WP_Widget_Recent_Posts' ); ?>
-
-					<?php if ( storefront_categorized_blog() ) : // Only show the widget if site has multiple categories. ?>
-					<div class="widget widget_categories">
-						<h2 class="widget-title"><?php _e( 'Most Used Categories', 'storefront' ); ?></h2>
-						<ul>
-						<?php
-							wp_list_categories( array(
-								'orderby'    => 'count',
-								'order'      => 'DESC',
-								'show_count' => 1,
-								'title_li'   => '',
-								'number'     => 10,
-							) );
-						?>
-						</ul>
-					</div><!-- .widget -->
-					<?php endif; ?>
+					<p><?php esc_html_e( 'Nothing was found at this location. Try searching, or check out the links below.', 'storefront' ); ?></p>
 
 					<?php
-						$archive_content = '<p>' . __( 'Try looking in the monthly archives.', 'storefront' ) . '</p>';
-						the_widget( 'WP_Widget_Archives', 'dropdown=1', "after_title=</h2>$archive_content" );
+					if ( is_woocommerce_activated() ) {
+						the_widget( 'WC_Widget_Product_Search' );
+					} else {
+						get_search_form();
+					}
 					?>
 
-					<?php the_widget( 'WP_Widget_Tag_Cloud' ); ?>
+					<?php
+					if ( is_woocommerce_activated() ) {
+
+						echo '<div class="fourohfour-columns-2">';
+
+							echo '<div class="col-1">';
+
+								storefront_promoted_products();
+
+							echo '</div>';
+
+							echo '<div class="col-2">';
+
+								echo '<h2>' . esc_html__( 'Product Categories', 'storefront' ) . '</h2>';
+
+								the_widget( 'WC_Widget_Product_Categories', array(
+															'count'		=> 1,
+														) );
+							echo '</div>';
+
+						echo '</div>';
+
+						echo '<h2>' . esc_html__( 'Popular Products', 'storefront' ) . '</h2>';
+
+						echo storefront_do_shortcode( 'best_selling_products', array(
+															'per_page' 	=> 4,
+															'columns'	=> 4,
+														) );
+					}
+					?>
 
 				</div><!-- .page-content -->
 			</section><!-- .error-404 -->