+++ inc/customizer/controls/radio-image.php
@@ -47,10 +47,10 @@ class Storefront_Custom_Radio_Image_Control extends WP_Customize_Control {
 		?>
 		<span class="customize-control-title">
 			<?php echo esc_attr( $this->label ); ?>
-			<?php if ( ! empty( $this->description ) ) : ?>
-				<span class="description customize-control-description"><?php echo esc_html( $this->description ); ?></span>
-			<?php endif; ?>
 		</span>
+		<?php if ( ! empty( $this->description ) ) : ?>
+			<span class="description customize-control-description"><?php echo esc_html( $this->description ); ?></span>
+		<?php endif; ?>
 		<div id="input_<?php echo esc_attr( $this->id ); ?>" class="image">
 			<?php foreach ( $this->choices as $value => $label ) : ?>
 				<input class="image-select" type="radio" value="<?php echo esc_attr( $value ); ?>" id="<?php echo $this->id . $value; ?>" name="<?php echo esc_attr( $name ); ?>" <?php $this->link(); checked( $this->value(), $value ); ?>>