+++ inc/functions/setup.php
@@ -40,7 +40,7 @@ if ( ! function_exists( 'storefront_setup' ) ) :
 		// wp-content/themes/child-theme-name/languages/it_IT.mo
 		load_theme_textdomain( 'storefront', get_stylesheet_directory() . '/languages' );
 
-		// wp-content/themes/theme-name/languages/it_IT.mo
+		// wp-content/themes/storefront/languages/it_IT.mo
 		load_theme_textdomain( 'storefront', get_template_directory() . '/languages' );
 
 		/**
@@ -143,7 +143,16 @@ function storefront_widgets_init() {
 function storefront_scripts() {
 	global $storefront_version;
 
-	wp_enqueue_style( 'storefront-style', get_stylesheet_uri(), '', $storefront_version );
+	if ( ! is_child_theme() ) {
+		// If a child theme's not loaded, just load the stylesheet as usual.
+		wp_enqueue_style( 'storefront-style', get_stylesheet_uri(), '', $storefront_version );
+	} else {
+		// If a child theme is loaded, load the parent theme styles as normal (includes rtl) as well as the child themes style.css file.
+		wp_enqueue_style( 'storefront-style', get_template_directory_uri() . '/style.css', '', $storefront_version );
+		wp_enqueue_style( 'storefront-child-style', get_stylesheet_uri(), '' );
+	}
+
+	wp_style_add_data( 'storefront-style', 'rtl', 'replace' );
 
 	wp_enqueue_script( 'storefront-navigation', get_template_directory_uri() . '/js/navigation.min.js', array(), '20120206', true );