+++ inc/woocommerce/template-tags.php
@@ -16,7 +16,7 @@ if ( ! function_exists( 'storefront_cart_link' ) ) {
 	function storefront_cart_link() {
 		?>
 			<a class="cart-contents" href="<?php echo esc_url( WC()->cart->get_cart_url() ); ?>" title="<?php _e( 'View your shopping cart', 'storefront' ); ?>">
-				<?php echo wp_kses_data( WC()->cart->get_cart_subtotal() ); ?> <span class="count"><?php echo wp_kses_data( sprintf( _n( '%d item', '%d items', WC()->cart->get_cart_contents_count(), 'storefront' ), WC()->cart->get_cart_contents_count() ) );?></span>
+				<span class="amount"><?php echo wp_kses_data( WC()->cart->get_cart_subtotal() ); ?></span> <span class="count"><?php echo wp_kses_data( sprintf( _n( '%d item', '%d items', WC()->cart->get_cart_contents_count(), 'storefront' ), WC()->cart->get_cart_contents_count() ) );?></span>
 			</a>
 		<?php
 	}
@@ -58,7 +58,9 @@ if ( ! function_exists( 'storefront_header_cart' ) ) {
 			<li class="<?php echo esc_attr( $class ); ?>">
 				<?php storefront_cart_link(); ?>
 			</li>
-			<?php the_widget( 'WC_Widget_Cart', 'title=' ); ?>
+			<li>
+				<?php the_widget( 'WC_Widget_Cart', 'title=' ); ?>
+			</li>
 		</ul>
 		<?php
 		}
@@ -103,7 +105,7 @@ function storefront_sorting_wrapper_close() {
  */
 function storefront_shop_messages() {
 	if ( ! is_checkout() ) {
-		echo wp_kses_post( do_shortcode_func( 'woocommerce_messages' ) );
+		echo wp_kses_post( storefront_do_shortcode( 'woocommerce_messages' ) );
 	}
 }
 
@@ -118,6 +120,50 @@ if ( ! function_exists( 'storefront_woocommerce_pagination' ) ) {
 	function storefront_woocommerce_pagination() {
 		if ( woocommerce_products_will_display() ) {
 			woocommerce_pagination();
+		}
+	}
+}
+
+/**
+ * Featured and On-Sale Products
+ * Check for featured products then on-sale products and use the appropiate shortcode. 
+ * If neither exist, it can fallback to show recently added products.
+ * @since  1.5.1
+ * @uses  is_woocommerce_activated()
+ * @uses  wc_get_featured_product_ids()
+ * @uses  wc_get_product_ids_on_sale()
+ * @uses  storefront_do_shortcode()
+ * @return void
+ */
+if ( ! function_exists( 'storefront_promoted_products' ) ) {
+	function storefront_promoted_products( $per_page = '2', $columns = '2', $recent_fallback = true ) {
+		if ( is_woocommerce_activated() ) { 
+
+			if ( wc_get_featured_product_ids() ) {
+
+				echo '<h2>' . esc_html__( 'Featured Products', 'storefront' ) . '</h2>';
+
+				echo storefront_do_shortcode( 'featured_products', array(
+											'per_page' 	=> $per_page,
+											'columns'	=> $columns,
+										) );
+			} elseif ( wc_get_product_ids_on_sale() ) {
+
+				echo '<h2>' . esc_html__( 'On Sale Now', 'storefront' ) . '</h2>';
+
+				echo storefront_do_shortcode( 'sale_products', array(
+											'per_page' 	=> $per_page,
+											'columns'	=> $columns,
+										) );
+			} elseif ( $recent_fallback ) {
+
+				echo '<h2>' . esc_html__( 'New In Store', 'storefront' ) . '</h2>';
+
+				echo storefront_do_shortcode( 'recent_products', array(
+											'per_page' 	=> $per_page,
+											'columns'	=> $columns,
+										) );
+			}
 		}
 	}
 }